requirements.txt
discord.py
git+https://github.com/Rapptz/discord-ext-menus
jishaku
asyncpg
